requirements/requirements_example.txt
scikit-learn
matplotlib
scipy
pandas
requests
torchvision==0.22.0
torchmetrics[image]